herokuDemo/requirements.txt
flask
line-bot-sdk
request
abort
boto3